# [Fix #77] Add new `Performance/BindCall` cop