@pocke 2020-01-07
Can you add a test case for `method.bind(obj).()` syntax?
I guess it works well so it only needs to add a test case, not modifying cop.

And I guess we can also check `Method#[]` method because it is the same as `Method#call`.
But I guess it is a rare case.
So probably we can skip checking `[]` in this pull request.