@koic 2020-01-07
> Can you add a test case for `method.bind(obj).()` syntax?

I have added support for this syntax to this cop. Thank you for letting me know about the syntax!